UPDATE LOGIC CHECKOUT SUCCESS 8/10

+++ clothingsystem/src/main/java/com/ecommerShop/clothingsystem/service/VNPayService.java
@@ -3,7 +3,7 @@ package com.ecommerShop.clothingsystem.service;
 import jakarta.servlet.http.HttpServletRequest;
 
 public interface VNPayService {
-    String createPaymentUrl(Long orderId, long amount, HttpServletRequest request);
+    String createPaymentUrl(Long orderId, double amount, HttpServletRequest request);
 
     int orderReturn(HttpServletRequest request);
 }
